projects/yolov8-poc/requirements.txt
# YOLOv8 Object Detection Dashboard Requirements
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

